WebGL: Add workaround for skinning with Emscripten.
So...we use integers for bone indices, and that makes them special. :)

On May 26, the emsdk WebGL 2.0 wrapper for the integer variant of
VertexAttribPointer was broken due to an incomplete renaming operation
from an emscripten contributor.

I have a pending upstream fix to the emscripten repo (#11742) but for
now we can workaround this with a simple monkeypatch.

Fixes #2856.